Utilizando variáveis globais
