+++ bytebank/src/main/kotlin/br/com/alura/bytebank/main.kt
@@ -1,3 +1,15 @@
+import br.com.alura.bytebank.modelo.Cliente
+import br.com.alura.bytebank.modelo.ContaCorrente
+import br.com.alura.bytebank.modelo.ContaPoupanca
+import br.com.alura.bytebank.modelo.totalContas
+
 fun main() {
+
+    val alex = Cliente(nome = "Alex", cpf = "", senha = 1)
+    val contaPoupanca = ContaPoupanca(titular = alex, numero = 1000)
+    val contaCorrente = ContaCorrente(titular = alex, numero = 1001)
+
     testaContasDiferentes()
+
+    println("Total de contas: $totalContas")
 }
