+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/Conta.kt
@@ -1,11 +1,19 @@
 package br.com.alura.bytebank.modelo
 
+var totalContas = 0
+    private set
+
 abstract class Conta(
     var titular: Cliente,
     val numero: Int
 ) {
     var saldo = 0.0
         protected set
+
+    init {
+        println("Criando conta!")
+        totalContas++
+    }
 
     init {
         println("Inicializando a classe de conta: $titular")
@@ -26,5 +34,34 @@ abstract class Conta(
             return true
         }
         return false
+    }
+}
+
+class ContaCorrente(
+    titular: Cliente,
+    numero: Int
+) : Conta(
+    titular = titular,
+    numero = numero
+) {
+    override fun sacar(valor: Double) {
+        val valorComTaxa = valor + 0.1
+        if (this.saldo >= valorComTaxa) {
+            this.saldo -= valorComTaxa
+        }
+    }
+}
+
+class ContaPoupanca(
+    titular: Cliente,
+    numero: Int
+) : Conta(
+    titular = titular,
+    numero = numero
+) {
+    override fun sacar(valor: Double) {
+        if (this.saldo >= valor) {
+            this.saldo -= valor
+        }
     }
 }
