--- bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/ContaCorrente.kt
@@ -1,16 +1,0 @@
-package br.com.alura.bytebank.modelo
-
-class ContaCorrente(
-    titular: Cliente,
-    numero: Int
-) : Conta(
-    titular = titular,
-    numero = numero
-) {
-    override fun sacar(valor: Double) {
-        val valorComTaxa = valor + 0.1
-        if (this.saldo >= valorComTaxa) {
-            this.saldo -= valorComTaxa
-        }
-    }
-}
